feat: DB접근 준비 및 Integer -> Long 변환

+++ src/main/java/com/deca/NoticeBoardWebSite/controller/PostDataController.java
@@ -51,7 +51,7 @@ public class PostDataController {
 
     @CrossOrigin(origins = "http://localhost:3000")
     @GetMapping("/getCount")
-    public Integer getPostCount(){
+    public Long getPostCount(){
         return postDataService.getPostCount();
     }
